out_request: emit per request latency in out req
